Setup Vision Processing Node - Incomplete
#2
