+++ msg/maze_info.msg
@@ -1,0 +1,4 @@
+Header header
+uint64 counter
+uint64 scale
+geometry_msgs/Pose maze_origin
